split dnsmap
